requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-multipart
requests
asyncpg
fitz